C++ win32: timeout+error handling
Windows doesn't seem to have a read() call with timeout, so we have to use
WaitForSingleObject and then read using ReadConsoleInput (ReadConsole doesn't
seem to work for this?). This is a more complex API than just reading
characters, but in the end it might be more useful.

Also in the build script, it ignores the subsystem argument and it doesn't seem
to affect anything, so just remove it.