Fix a bug with initial newlines in new files
This was being fooled by picking the "newline" from the end of the first line of
a new file (before any newlines have been inserted). It didn't check what the
string consisted of, so it just plonked it on the next line etc. leaving no eol
anywhere in the file. It was ok with any file that already had newlines in it.