Fix bug with cursor positioning in empty file buffers
The problem was that we expect every line to end in a newline character,
which the cursor occupies when it's at the end, where you type,
essentially you're always inserting text in between the text of the line
and the newline character.

In empty files, we never insert that newline for the row(s) inserted. It
looks like we'd have the same problem if we typed at the end of the
file? Maybe not... Anyway, it's fixed now, typing in empty file buffers
works as normal.